Pathfinder integration cleanup (#302)
A follow up from #287 to clean things up.
