+++ Cargo.toml
@@ -33,6 +33,7 @@ cairo-vm = { version = "=1.0.0-rc5", features = ["extensive_hints", "cairo-1-hin
 clap = { version = "4.5.4", features = ["derive"] }
 env_logger = "0.11.3"
 futures = "0.3.30"
+futures-core = "0.3.30"
 futures-util = "0.3.30"
 heck = "0.4.1"
 hex = "0.4.3"
@@ -53,6 +54,7 @@ serde = { version = "1.0.188", features = ["derive"] }
 serde_json = { version = "1.0.105", features = ["arbitrary_precision"] }
 serde_with = "3.3.0"
 serde_yaml = "0.9.25"
+starknet = "0.11.0"
 starknet_api = { version = "=0.10", features = ["testing"] }
 starknet-core = "0.11.1"
 starknet-crypto = "0.6.2"
