+++ crates/bin/prove_block/src/main.rs
@@ -1,6 +1,5 @@
 use std::collections::{HashMap, HashSet};
 use std::error::Error;
-use std::future::Future;
 
 use blockifier::block::{BlockInfo, GasPrices};
 use blockifier::context::{BlockContext, ChainInfo, FeeTokenAddresses};
@@ -11,12 +10,8 @@ use cairo_vm::vm::errors::cairo_run_errors::CairoRunError::VmException;
 use cairo_vm::Felt252;
 use clap::Parser;
 use reexecute::{reexecute_transactions_with_blockifier, RpcStateReader};
-use serde::de::DeserializeOwned;
-use serde::Deserialize;
-use serde_json::json;
-use starknet::core::types::{
-    BlockId, BlockWithTxs, MaybePendingBlockWithTxs, MaybePendingStateUpdate, StateUpdate, StorageEntry,
-};
+use rpc_utils::{get_storage_proofs, RpcStorage, TrieNode};
+use starknet::core::types::{BlockId, BlockWithTxs, MaybePendingBlockWithTxs, MaybePendingStateUpdate};
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::{JsonRpcClient, Provider, Url};
 use starknet_api::block::{BlockNumber, BlockTimestamp};
@@ -39,17 +34,17 @@ use starknet_os::starknet::starknet_storage::{CommitmentInfo, OsSingleStarknetSt
 use starknet_os::starkware_utils::commitment_tree::base_types::{Height, Length, NodePath, TreeIndex};
 use starknet_os::starkware_utils::commitment_tree::binary_fact_tree::BinaryFactTree;
 use starknet_os::starkware_utils::commitment_tree::patricia_tree::nodes::{BinaryNodeFact, EdgeNodeFact};
-use starknet_os::storage::cached_storage::CachedStorage;
-use starknet_os::storage::storage::{Fact, FactFetchingContext, Storage, StorageError};
+use starknet_os::storage::storage::{Fact, FactFetchingContext};
 use starknet_os::utils::felt_vm2api;
 use starknet_os::{config, run_os};
 use starknet_os_types::casm_contract_class::GenericCasmContractClass;
-use starknet_types_core::felt::Felt;
 use types::starknet_rs_to_blockifier;
 
+use crate::rpc_utils::CachedRpcStorage;
 use crate::types::starknet_rs_tx_to_internal_tx;
 
 mod reexecute;
+mod rpc_utils;
 mod types;
 
 #[derive(Parser, Debug)]
@@ -63,172 +58,9 @@ struct Args {
     rpc_provider: String,
 }
 
-fn jsonrpc_request(method: &str, params: serde_json::Value) -> serde_json::Value {
-    json!({
-        "jsonrpc": "2.0",
-        "id": "0",
-        "method": method,
-        "params": params,
-    })
-}
-
-async fn post_jsonrpc_request<T: DeserializeOwned>(
-    client: &reqwest::Client,
-    rpc_provider: &str,
-    method: &str,
-    params: serde_json::Value,
-) -> Result<T, reqwest::Error> {
-    let request = jsonrpc_request(method, params);
-    let response = client.post(format!("{}/rpc/v0_7", rpc_provider)).json(&request).send().await?;
-
-    #[derive(Deserialize)]
-    struct TransactionReceiptResponse<T> {
-        result: T,
-    }
-
-    println!("Response status: {}", response.status());
-    let response_text = response.text().await?;
-    let response: TransactionReceiptResponse<T> =
-        serde_json::from_str(&response_text).unwrap_or_else(|_| panic!("Error: {}", response_text));
-    Ok(response.result)
-}
-
-#[derive(Clone, Deserialize)]
-struct EdgePath {
-    len: u64,
-    value: Felt252,
-}
-
-#[derive(Clone, Deserialize)]
-enum TrieNode {
-    #[serde(rename = "binary")]
-    Binary { left: Felt252, right: Felt252 },
-    #[serde(rename = "edge")]
-    Edge { child: Felt252, path: EdgePath },
-}
-
-#[derive(Deserialize)]
-pub struct ContractData {
-    /// The proofs associated with the queried storage values
-    storage_proofs: Vec<Vec<TrieNode>>,
-}
-
-#[derive(Deserialize)]
-struct PathfinderProof {
-    class_commitment: Felt252,
-    state_commitment: Felt252,
-    contract_proof: Vec<TrieNode>,
-    contract_data: Option<ContractData>,
-}
-
-async fn pathfinder_get_proof(
-    client: &reqwest::Client,
-    rpc_provider: &str,
-    block_number: u64,
-    contract_address: Felt,
-    keys: &[Felt],
-) -> Result<PathfinderProof, reqwest::Error> {
-    post_jsonrpc_request(
-        client,
-        rpc_provider,
-        "pathfinder_getProof",
-        json!({ "block_id": { "block_number": block_number }, "contract_address": contract_address, "keys": keys }),
-    )
-    .await
-}
-
-async fn get_storage_proofs(
-    client: &reqwest::Client,
-    rpc_provider: &str,
-    block_number: u64,
-    state_update: &StateUpdate,
-) -> Result<HashMap<Felt, PathfinderProof>, reqwest::Error> {
-    let mut storage_changes_by_contract: HashMap<Felt, Vec<StorageEntry>> = HashMap::new();
-
-    for diff_item in &state_update.state_diff.storage_diffs {
-        storage_changes_by_contract.entry(diff_item.address).or_default().extend_from_slice(&diff_item.storage_entries);
-    }
-
-    let mut storage_proofs = HashMap::new();
-
-    for (contract_address, storage_changes) in storage_changes_by_contract {
-        let keys: Vec<_> = storage_changes.iter().map(|change| change.key).collect();
-
-        // The endpoint is limited to 100 keys at most per call
-        const MAX_KEYS: usize = 100;
-        let mut chunked_storage_proofs = Vec::new();
-        for keys_chunk in keys.chunks(MAX_KEYS) {
-            chunked_storage_proofs
-                .push(pathfinder_get_proof(client, rpc_provider, block_number, contract_address, keys_chunk).await?);
-        }
-        let storage_proof = merge_chunked_storage_proofs(chunked_storage_proofs);
-
-        storage_proofs.insert(contract_address, storage_proof);
-    }
-
-    Ok(storage_proofs)
-}
-
-fn merge_chunked_storage_proofs(proofs: Vec<PathfinderProof>) -> PathfinderProof {
-    let class_commitment = proofs[0].class_commitment;
-    let state_commitment = proofs[0].state_commitment;
-    let contract_proof = proofs[0].contract_proof.clone();
-
-    let contract_data = {
-        let mut contract_data: Option<ContractData> = None;
-
-        for proof in proofs {
-            if let Some(data) = proof.contract_data {
-                if let Some(contract_data) = contract_data.as_mut() {
-                    contract_data.storage_proofs.extend(data.storage_proofs);
-                } else {
-                    contract_data = Some(data);
-                }
-            }
-        }
-
-        contract_data
-    };
-
-    PathfinderProof { class_commitment, state_commitment, contract_proof, contract_data }
-}
-
 fn felt_to_u128(felt: &starknet_types_core::felt::Felt) -> u128 {
     let digits = felt.to_be_digits();
     ((digits[2] as u128) << 64) + digits[3] as u128
-}
-
-#[derive(Clone)]
-struct RpcStorage {
-    // provider: JsonRpcClient<HttpTransport>,
-    // client: reqwest::Client,
-    // provider: String,
-    // block_number: u64,
-}
-
-impl RpcStorage {
-    // pub fn new(provider: JsonRpcClient<HttpTransport>) -> Self {
-    pub fn new(_client: reqwest::Client, _provider: String, _block_number: u64) -> Self {
-        Self {}
-    }
-}
-
-type CachedRpcStorage = CachedStorage<RpcStorage>;
-
-impl Storage for RpcStorage {
-    async fn set_value(&mut self, _key: Vec<u8>, _value: Vec<u8>) -> Result<(), StorageError> {
-        // log::warn!("RpcStorage ignoring attempt to write storage - {:?}: {:?}", key, value);
-        Ok(())
-    }
-
-    fn get_value(&self, key: &[u8]) -> impl Future<Output = Result<Option<Vec<u8>>, StorageError>> + Send {
-        log::info!("RpcStorage get_value, key-len: {}, key: {:?}", key.len(), key);
-        async {
-            // let response = pathfinder_get_proof(&self.client, &self.provider, self.block_number,
-            // contract_address, keys_chunk).await .map_err(|_| StorageError::ContentNotFound);
-            Ok(Some(Default::default()))
-        }
-    }
 }
 
 async fn build_block_context(chain_id: String, block: &BlockWithTxs) -> Result<BlockContext, reqwest::Error> {
@@ -377,7 +209,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let layout = LayoutName::starknet_with_keccak;
 
     let provider_url = format!("{}/rpc/v0_7", args.rpc_provider);
-    println!("provider url: {}", provider_url);
+    log::info!("provider url: {}", provider_url);
     let provider = JsonRpcClient::new(HttpTransport::new(
         Url::parse(provider_url.as_str()).expect("Could not parse provider url"),
     ));
@@ -423,7 +255,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     // initialize storage. We use a CachedStorage with a RcpStorage as the main storage, meaning
     // that a DictStorage serves as the cache layer and we will use Pathfinder RPC for cache misses
-    let rpc_storage = RpcStorage::new(pathfinder_client, provider_url, block_number);
+    let rpc_storage = RpcStorage::new();
     let cached_storage = CachedRpcStorage::new(Default::default(), rpc_storage);
 
     // TODO: nasty clone, the conversion fns don't take references
@@ -625,7 +457,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
             log::debug!("querying class for contract_address {}, hash: {}", contract_address, class_hash);
             if let Ok(contract_class) = provider.get_class(BlockId::Number(block_number), class_hash).await {
-                log::trace!("contract class for {}: {:?}", contract_address, contract_class);
+                log::debug!("contract class for {}: {:?}", contract_address, contract_class);
 
                 match contract_class {
                     starknet::core::types::ContractClass::Sierra(flattened_sierra_cc) => {
@@ -649,7 +481,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                             .unwrap();
                         let casm_cc: GenericCasmContractClass = casm_cc.into();
 
-                        log::error!("class_hash (from RPC): {}", class_hash);
+                        log::debug!("class_hash (from RPC): {}", class_hash);
 
                         // TODO: it seems that this ends up computing the wrong class hash...
                         write_class_facts(
@@ -730,7 +562,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
             }
         }
         Err(_) => {
-            println!("exception:\n{:#?}", result);
+            log::error!("exception:\n{:#?}", result);
         }
         _ => {}
     }
